refactor: FileBrowserPaneView のダイアログ表示を FileBrowserDialogs ヘルパーに分離 (#156)
ダイアログ・ピッカー表示を View コードビハインドから internal sealed の
FileBrowserDialogs へ移動し、FileBrowserPaneView.xaml.cs を 1,921→1,515 行
（-406行）に削減。

- 競合解決(Move/Copy)・テキスト入力・確認・メッセージ・各種操作エラー
  (作成/リネーム/Move/Copy/削除)・フォルダピッカー(HWND Interop含む)・
  EnsureXamlRootAsync を FileBrowserDialogs.cs へ移動
- View は XamlRoot 供給元(RootGrid)と HostWindow アクセサをコンストラクタで
  渡し、_dialogs.ShowXxxAsync(...) 経由で呼び出す。VM へ渡す競合解決コール
  バック型(Func<string,bool,Task<ConflictResolution>>)は不変
- リソースキー以外ほぼ同一だった ShowMove/CopyConflictDialogAsync を、
  接頭辞をパラメータ化した単一実装 ShowConflictAsync に統合
- 操作エラーダイアログ群は分岐表が操作種別ごとに異なるため統合せず移動のみ
  （汎用 DialogService との統合可否は #156 のフォローアップに記録）

ダイアログ表示は View の正当な責務のため挙動は不変。build/test(576件)・
dotnet format(Release,x64)すべてパス。
